Include periodic jobs execution for the collection
This collection is not pinning any of the requirements
versions, if there is an update i.e. in the pytest or
ansible lint packages the linters, integration or unit tests
might start to fail for unrelated reasons.
This commit add periodic executions in the GH actions workflows
and includes a badge per job in the README.md file.